pyproject.toml
[tool.poetry]
name = "lakehouse-tools"
version = "0.1.0"
description = "Tools for Data Lakehouses"
authors = ["Alexander Wagner <first.last@domain.com>"]
readme = "README.md"
packages = [{include = "lakehouse_tools"}]

[tool.poetry.dependencies]
python = "^3.10"
delta-spark = "^2.3.0"
ipykernel = "^6.25.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3"
chispa = "^0.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"